Make StationCrowding inherit from TrainCrowding
