+++ lib/blocktrain.rb
@@ -8,8 +8,9 @@ require 'singleton'
 require_relative 'blocktrain/client'
 require_relative 'blocktrain/query'
 require_relative 'blocktrain/count'
-require_relative 'blocktrain/atp_query'
 require_relative 'blocktrain/train_crowding'
+require_relative 'blocktrain/atp_query'
+require_relative 'blocktrain/station_crowding'
 require_relative 'blocktrain/aggregation'
 require_relative 'blocktrain/paginated_query'
 require_relative 'blocktrain/aggregations/histogram_aggregation'
