+++ spec/blocktrain/station_crowding_spec.rb
@@ -1,10 +1,10 @@
 module Blocktrain
-  describe TrainCrowding do
-    it 'sends the right paramesters to ATPQuery' do
+  describe StationCrowding do
+    it 'sends the right paramesters to ATPQuery', :vcr do
       from = '1974-06-14T00:00:00+00:00'
       to = '1974-06-15T00:00:00+00:00'
 
-      expect(ATPQuery).to receive(:new).with(from: from, to: to, station: :euston, direction: :southbound)
+      expect(ATPQuery).to receive(:new).with(from: from, to: to, station: :euston, direction: :southbound).and_call_original
 
       tc = described_class.new Time.parse(to), :euston, :southbound
     end
@@ -23,19 +23,17 @@ module Blocktrain
     it 'has results', :vcr do
       tc = described_class.new Time.parse('2015-09-17T17:00:00.000Z'), :euston, :southbound
 
-      expect(tc.results).to eq [
-        [
-          {
-            "number"=>0,
-            "timeStamp"=>"2015-09-17T16:02:22.521Z"
-          },
-          {
-            "CAR_A"=>25.333333333333332,
-            "CAR_B"=>33.83582089552239,
-            "CAR_C"=>26.193548387096776,
-            "CAR_D"=>10.428571428571429
-          }
-        ]
+      expect(tc.results).to include [
+        {
+          "number"=>0,
+          "timeStamp"=>"2015-09-17T16:02:22.521Z"
+        },
+        {
+          "CAR_A"=>25.333333333333332,
+          "CAR_B"=>33.83582089552239,
+          "CAR_C"=>26.193548387096776,
+          "CAR_D"=>10.428571428571429
+        }
       ]
     end
